28 Juni 2021

+++ app/src/main/java/com/misit/abpenergy/Sarpras/SaranaResponse/PenumpangListModel.kt
@@ -1,10 +1,11 @@
 package com.misit.abpenergy.Sarpras.SaranaResponse
 
+import android.util.Log
 import com.google.gson.annotations.SerializedName
 
 data class PenumpangListModel(
     @field:SerializedName("id")
-    var id: Int? = null,
+    var id: Long? = null,
     @field:SerializedName("nik")
     var nik: String? = null,
     @field:SerializedName("nama")
